separate component & remove unused component code
