+++ src/layout/BasicDrawerHeader.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import {styled} from "@mui/material/styles";
+
+const DrawerHeader = styled('div')(({ theme }) => ({
+    display: 'flex',
+    alignItems: 'center',
+    padding: theme.spacing(0, 1),
+    // necessary for content to be below app bar
+    ...theme.mixins.toolbar,
+    justifyContent: 'flex-end',
+}));
+
+export default DrawerHeader;
