+++ src/layout/PageLayout.jsx
@@ -1,83 +1,21 @@
 import * as React from 'react';
-import {createTheme, styled, ThemeProvider, useTheme} from '@mui/material/styles';
+import {createTheme, ThemeProvider} from '@mui/material/styles';
 import Box from '@mui/material/Box';
-import Drawer from '@mui/material/Drawer';
 import CssBaseline from '@mui/material/CssBaseline';
-import MuiAppBar from '@mui/material/AppBar';
 import Toolbar from '@mui/material/Toolbar';
-import List from '@mui/material/List';
-import Typography from '@mui/material/Typography';
-import Divider from '@mui/material/Divider';
 import IconButton from '@mui/material/IconButton';
 import MenuIcon from '@mui/icons-material/Menu';
-import ChevronLeftIcon from '@mui/icons-material/ChevronLeft';
-import ChevronRightIcon from '@mui/icons-material/ChevronRight';
-import ListItem from '@mui/material/ListItem';
-import ListItemButton from '@mui/material/ListItemButton';
-import ListItemIcon from '@mui/material/ListItemIcon';
-import ListItemText from '@mui/material/ListItemText';
-import InboxIcon from '@mui/icons-material/MoveToInbox';
-import MailIcon from '@mui/icons-material/Mail';
-import MenuList from "../page/MenuList";
-import MenuDetail from "../page/MenuDetail";
 import {Button, Fab} from "@mui/material";
-import EditIcon from "@mui/icons-material/Edit";
-import ReviewCreate from "../page/ReviewCreate";
 import {grey} from "@mui/material/colors";
 import {useNavigate} from "react-router-dom";
-
+import BasicDrawer from "./BasicDrawer";
+import AppBar from "./BasicAppBar";
+import Main from "./BasicMain";
+import DrawerHeader from "./BasicDrawerHeader";
 
 const drawerWidth = 240;
 
-const Main = styled('main', { shouldForwardProp: (prop) => prop !== 'open' })(
-    ({ theme, open }) => ({
-        flexGrow: 1,
-        padding: theme.spacing(3),
-        transition: theme.transitions.create('margin', {
-            easing: theme.transitions.easing.sharp,
-            duration: theme.transitions.duration.leavingScreen,
-        }),
-        marginLeft: `-${drawerWidth}px`,
-        ...(open && {
-            transition: theme.transitions.create('margin', {
-                easing: theme.transitions.easing.easeOut,
-                duration: theme.transitions.duration.enteringScreen,
-            }),
-            marginLeft: 0,
-        }),
-    }),
-);
-
-const AppBar = styled(MuiAppBar, {
-    shouldForwardProp: (prop) => prop !== 'open',
-})(({ theme, open }) => ({
-    backgroundColor: theme.palette.primary.main,
-    transition: theme.transitions.create(['margin', 'width'], {
-        easing: theme.transitions.easing.sharp,
-        duration: theme.transitions.duration.leavingScreen,
-    }),
-    ...(open && {
-        width: `calc(100% - ${drawerWidth}px)`,
-        marginLeft: `${drawerWidth}px`,
-        transition: theme.transitions.create(['margin', 'width'], {
-            easing: theme.transitions.easing.easeOut,
-            duration: theme.transitions.duration.enteringScreen,
-        }),
-    }),
-}));
-
-const DrawerHeader = styled('div')(({ theme }) => ({
-    display: 'flex',
-    alignItems: 'center',
-    padding: theme.spacing(0, 1),
-    // necessary for content to be below app bar
-    ...theme.mixins.toolbar,
-    justifyContent: 'flex-end',
-}));
-
-
 const PageLayout = ({children, isDetailPage}) => {
-    // const theme = useTheme();
     const myTheme = createTheme({
         palette: {
             primary: {
@@ -107,16 +45,16 @@ const PageLayout = ({children, isDetailPage}) => {
 
     const navigate = useNavigate();
 
-    const handleClickMenuList = () => {
+    const handleToolBarButtonClick = () => {
         navigate('/list');
-    };
+    }
 
     return (
         <ThemeProvider theme={myTheme}>
             <Box color="primary">
                 <Box sx={{ display: 'flex'}}>
                     <CssBaseline />
-                    <AppBar position="fixed" open={open}>
+                    <AppBar position="fixed" open={open} drawerWidth={drawerWidth}>
                         <Toolbar>
                             <IconButton
                                 color="inherit"
@@ -127,41 +65,23 @@ const PageLayout = ({children, isDetailPage}) => {
                             >
                                 <MenuIcon />
                             </IconButton>
-                            <Button variant="text" sx={{color: '#fff', fontSize: '1.5em'}} noWrap >
+                            <Button
+                                variant="text"
+                                sx={{color: '#fff', fontSize: '1.5em'}}
+                                noWrap
+                                onClick={handleToolBarButtonClick}
+                            >
                                 한성대 학식 리뷰
                             </Button>
                         </Toolbar>
                     </AppBar>
-                    <Drawer
-                        sx={{
-                            width: drawerWidth,
-                            flexShrink: 0,
-                            '& .MuiDrawer-paper': {
-                                width: drawerWidth,
-                                boxSizing: 'border-box',
-                            },
-                        }}
-                        variant="persistent"
-                        anchor="left"
+                    <BasicDrawer
+                        handleDrawerClose={handleDrawerClose}
                         open={open}
-                    >
-                        <DrawerHeader>
-                            <IconButton onClick={handleDrawerClose}>
-                                {myTheme.direction === 'ltr' ? <ChevronLeftIcon /> : <ChevronRightIcon />}
-                            </IconButton>
-                        </DrawerHeader>
-                        <Divider />
-                        <List>
-                            {['덮밥류 & 비빔밥', '면 & 찌개 & 김밥'].map((text, index) => (
-                                <ListItem key={text} disablePadding>
-                                    <ListItemButton sx={{textAlign: 'center', padding: '0.5em'}} onClick={handleClickMenuList}>
-                                        <ListItemText primary={text} />
-                                    </ListItemButton>
-                                </ListItem>
-                            ))}
-                        </List>
-                    </Drawer>
-                    <Main open={open}>
+                        drawerWidth={drawerWidth}
+                        myTheme={myTheme}
+                    />
+                    <Main open={open} drawerWidth={drawerWidth}>
                         <DrawerHeader />
                         {children}
                     </Main>
